refactor: move equipment and feat selection logic from ui to core
Extract equipped display data builders and feat proficiency
accumulation so UI stays presentation-only.
