+++ core/mechanics/proficiencies.py
@@ -357,6 +357,16 @@ def get_proficiency_choices(
     return choices
 
 
+def apply_picked_tools_to_character(
+    character: Character,
+    picked: list[str],
+) -> None:
+    """Добавить выбранные инструменты к владениям персонажа."""
+    character.tool_proficiencies = merge_proficiency_tokens(
+        character.tool_proficiencies, picked
+    )
+
+
 def apply_subclass_proficiencies_to_character(
     character: Character,
     subclass_id: str,
@@ -379,6 +389,7 @@ def apply_subclass_proficiencies_to_character(
 
 __all__ = [
     "ProficiencyChoice",
+    "apply_picked_tools_to_character",
     "merge_proficiency_tokens",
     # Checks
     "has_weapon_proficiency",
